# Core/Bones.Api.Client/AutoGen/Models/GetAssetByIdResponse.cs
// <auto-generated/>
#pragma warning disable CS0618
#nullable enable
using ReQuesty.Runtime.Abstractions.Serialization;
using ReQuesty.Runtime.Extensions;
using System.Collections.Generic;
using System.IO;
using System;
namespace Bones.Api.Client.AutoGen.Models
{
    /// <summary>
    /// Response for the GetAssetById endpoint
    /// </summary>
    [global::System.CodeDom.Compiler.GeneratedCode("ReQuesty", "0.0.1")]
    public partial class GetAssetByIdResponse : IParsable
    {
        /// <summary>ID of the asset</summary>
        public Guid AssetId { get; set; } = default!;
        /// <summary>The time this asset was created</summary>
        public DateTimeOffset CreateDateTime { get; set; } = default!;
        /// <summary>The current version number for this asset</summary>
        public int CurrentVersion { get; set; } = default!;
        /// <summary>The friendly ID of the asset</summary>
        public string FriendlyId { get; set; } = default!;
        /// <summary>The ID of the  item this asset is</summary>
        public Guid ItemId { get; set; } = default!;
        /// <summary>The values for this asset</summary>
        public List<global::Bones.Api.Client.AutoGen.Models.ItemValueDisplayModel> ItemValues { get; set; } = default!;
        /// <summary>The ID of the latest version of the  item this asset is</summary>
        public Guid LatestItemVersionId { get; set; } = default!;
        /// <summary>The time the latest version of this asset was created</summary>
        public DateTimeOffset LatestVersionCreateDateTime { get; set; } = default!;
        /// <summary>The ID of the layout this asset uses</summary>
        public Guid LayoutId { get; set; } = default!;
        /// <summary>The ID of the project this asset belongs to</summary>
        public Guid ProjectId { get; set; } = default!;
        /// <summary>The title of the asset</summary>
        public string Title { get; set; } = default!;
        /// <summary>
        /// Creates a new instance of the appropriate class based on discriminator value
        /// </summary>
        /// <returns>A <see cref="global::Bones.Api.Client.AutoGen.Models.GetAssetByIdResponse"/></returns>
        /// <param name="parseNode">The parse node to use to read the discriminator value and create the object</param>
        public static global::Bones.Api.Client.AutoGen.Models.GetAssetByIdResponse CreateFromDiscriminatorValue(IParseNode parseNode)
        {
            _ = parseNode ?? throw new ArgumentNullException(nameof(parseNode));
            return new global::Bones.Api.Client.AutoGen.Models.GetAssetByIdResponse();
        }
        /// <summary>
        /// The deserialization information for the current model
        /// </summary>
        /// <returns>A IDictionary&lt;string, Action&lt;IParseNode&gt;&gt;</returns>
        public virtual IDictionary<string, Action<IParseNode>> GetFieldDeserializers()
        {
            return new Dictionary<string, Action<IParseNode>>
            {
                { "assetId", n => { AssetId = n.GetGuidValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'AssetId'"); } },
                { "createDateTime", n => { CreateDateTime = n.GetDateTimeOffsetValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'CreateDateTime'"); } },
                { "currentVersion", n => { CurrentVersion = n.GetIntValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'CurrentVersion'"); } },
                { "friendlyId", n => { FriendlyId = n.GetStringValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'FriendlyId'"); } },
                { "itemId", n => { ItemId = n.GetGuidValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'ItemId'"); } },
                { "itemValues", n => { ItemValues = n.GetCollectionOfObjectValues<global::Bones.Api.Client.AutoGen.Models.ItemValueDisplayModel>(global::Bones.Api.Client.AutoGen.Models.ItemValueDisplayModel.CreateFromDiscriminatorValue)?.AsList() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'ItemValues'"); } },
                { "latestItemVersionId", n => { LatestItemVersionId = n.GetGuidValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'LatestItemVersionId'"); } },
                { "latestVersionCreateDateTime", n => { LatestVersionCreateDateTime = n.GetDateTimeOffsetValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'LatestVersionCreateDateTime'"); } },
                { "layoutId", n => { LayoutId = n.GetGuidValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'LayoutId'"); } },
                { "projectId", n => { ProjectId = n.GetGuidValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'ProjectId'"); } },
                { "title", n => { Title = n.GetStringValue() ?? throw new NullReferenceException("Unexpected null value for non-nullable property: 'Title'"); } },
            };
        }
        /// <summary>
        /// Serializes information the current object
        /// </summary>
        /// <param name="writer">Serialization writer to use to serialize this model</param>
        public virtual void Serialize(ISerializationWriter writer)
        {
            _ = writer ?? throw new ArgumentNullException(nameof(writer));
            writer.WriteGuidValue("assetId", AssetId);
            writer.WriteDateTimeOffsetValue("createDateTime", CreateDateTime);
            writer.WriteIntValue("currentVersion", CurrentVersion);
            writer.WriteStringValue("friendlyId", FriendlyId);
            writer.WriteGuidValue("itemId", ItemId);
            writer.WriteCollectionOfObjectValues<global::Bones.Api.Client.AutoGen.Models.ItemValueDisplayModel>("itemValues", ItemValues);
            writer.WriteGuidValue("latestItemVersionId", LatestItemVersionId);
            writer.WriteDateTimeOffsetValue("latestVersionCreateDateTime", LatestVersionCreateDateTime);
            writer.WriteGuidValue("layoutId", LayoutId);
            writer.WriteGuidValue("projectId", ProjectId);
            writer.WriteStringValue("title", Title);
        }
    }
}
#nullable restore
#pragma warning restore CS0618