# Services/Bones.Api/OpenApi/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Bones API",
    "description": "Its an API and it does stuff",
    "contact": {
      "name": "GitHub Issues",
      "url": "https://github.com/CorruptComputer/Bones/issues"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/CorruptComputer/Bones/blob/develop/LICENSE"
    },
    "version": "develop"
  },
  "paths": {
...
    "/Asset/Layout/{ItemLayoutId}/dashboard": {
      "get": {
        "tags": [
          "Asset"
        ],
        "summary": "Gets the dashboard for an asset layout, which includes the layout and all assets that use it.",
        "operationId": "GetAssetLayoutDashboardAsync",
        "parameters": [
          {
            "name": "ItemLayoutId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAssetLayoutDashboardResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Asset/{AssetId}": {
      "get": {
        "tags": [
          "Asset"
        ],
        "summary": "Gets an asset by its ID",
        "operationId": "GetAssetByIdAsync",
        "parameters": [
          {
            "name": "AssetId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAssetByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/Asset/action/create": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "Creates a new asset",
        "operationId": "CreateAssetActionAsync",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetAction"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/Asset/action/delete": {
      "post": {
        "tags": [
          "Asset"
        ],
        "summary": "Deletes an asset",
        "operationId": "DeleteAssetActionAsync",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAssetAction"
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetActionBase": {
        "required": [
          "actionDateTime"
        ],
        "type": "object",
        "properties": {
          "actionDateTime": {
            "type": "string",
            "description": "The timestamp of when this action was performed",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "An action that can be performed on an asset"
      },
      "AssetActionResponse": {
        "required": [
          "assetCurrentVersionId",
          "assetId"
        ],
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "description": "The unique identifier of the asset action.",
            "format": "uuid"
          },
          "assetCurrentVersionId": {
            "type": "string",
            "description": "The unique identifier of the current version of the asset.",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "Response model for asset actions."
      },
...
      "CreateAssetAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetActionBase"
          },
          {
            "required": [
              "assetLayoutId",
              "fieldValues",
              "title"
            ],
            "type": "object",
            "properties": {
              "assetLayoutId": {
                "type": "string",
                "description": "The ID of the asset layout (not version, automatically uses the current version)",
                "format": "uuid"
              },
              "title": {
                "type": "string",
                "description": "The title of the asset"
              },
              "fieldValues": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ItemValueModel"
                },
                "description": "The fields of the asset"
              }
            },
            "additionalProperties": false
          }
        ],
        "description": "Action to create a new asset."
      },
...
      "DashboardAssetModel": {
        "required": [
          "friendlyId",
          "id",
          "latestVersionCreateDateTime",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The ID of the work item",
            "format": "uuid"
          },
          "friendlyId": {
            "type": "string",
            "description": "The friendly ID of the work item"
          },
          "title": {
            "type": "string",
            "description": "The title of the work item"
          },
          "latestVersionCreateDateTime": {
            "type": "string",
            "description": "The date and time the latest version of the asset was created",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Model for a work item in the dashboard"
      },
...
      "DeleteAssetAction": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetActionBase"
          },
          {
            "required": [
              "assetId"
            ],
            "type": "object",
            "properties": {
              "assetId": {
                "type": "string",
                "description": "The ID of the asset to perform the action on",
                "format": "uuid"
              }
            },
            "additionalProperties": false
          }
        ],
        "description": "Action to delete an asset."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StringStringListKeyValuePair"
            },
            "description": "The errors that occurred, with the key being either the input that was invalid and the list of reasons it was invalid, or 'server' with the list of server errors.",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "description": "The error message, if any.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "The response body is empty, this is a workaround for the limitations of the API client."
      },
      "FieldType": {
        "enum": [
          "textField",
          "textBox",
          "integer",
          "decimal",
          "boolean",
          "dateTime",
          "valueList",
          "geoLocation"
        ],
        "type": "string"
      },
...
      "GetAssetByIdResponse": {
        "required": [
          "assetId",
          "createDateTime",
          "currentVersion",
          "friendlyId",
          "itemId",
          "itemValues",
          "latestItemVersionId",
          "latestVersionCreateDateTime",
          "layoutId",
          "projectId",
          "title"
        ],
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "description": "ID of the asset",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "The title of the asset"
          },
          "layoutId": {
            "type": "string",
            "description": "The ID of the layout this asset uses",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "description": "The ID of the project this asset belongs to",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "description": "The ID of the  item this asset is",
            "format": "uuid"
          },
          "latestItemVersionId": {
            "type": "string",
            "description": "The ID of the latest version of the  item this asset is",
            "format": "uuid"
          },
          "currentVersion": {
            "type": "integer",
            "description": "The current version number for this asset",
            "format": "int32"
          },
          "friendlyId": {
            "type": "string",
            "description": "The friendly ID of the asset"
          },
          "createDateTime": {
            "type": "string",
            "description": "The time this asset was created",
            "format": "date-time"
          },
          "latestVersionCreateDateTime": {
            "type": "string",
            "description": "The time the latest version of this asset was created",
            "format": "date-time"
          },
          "itemValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemValueDisplayModel"
            },
            "description": "The values for this asset"
          }
        },
        "additionalProperties": false,
        "description": "Response for the GetAssetById endpoint"
      },
      "GetAssetLayoutDashboardResponse": {
        "required": [
          "assets",
          "layoutName"
        ],
        "type": "object",
        "properties": {
          "layoutName": {
            "type": "string",
            "description": "The name of the layout"
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardAssetModel"
            },
            "description": "The assets in this layout, ordered by date added from oldest to newest"
          }
        },
        "additionalProperties": false,
        "description": "Response for the GetAssetLayoutDashboardAsync endpoint"
      },
...
      "ItemValueDisplayModel": {
        "required": [
          "fieldVersionId",
          "isRequired",
          "name",
          "orderNumber",
          "valueType"
        ],
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "integer",
            "description": "The order that this field should be displayed in",
            "format": "int32"
          },
          "fieldVersionId": {
            "type": "string",
            "description": "The ID of the field version this value is for",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "The name of the field"
          },
          "valueType": {
            "$ref": "#/components/schemas/FieldType"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Is this field required?"
          },
          "canBeNegative": {
            "type": "boolean",
            "description": "If the field is a number, can it be negative?",
            "nullable": true
          },
          "possibleValues": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "If this field is a value list, the possible values for the field",
            "nullable": true
          },
          "value": {
            "type": "string",
            "description": "The value of the field",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Model for the values of an item"
      },
      "ItemValueModel": {
        "required": [
          "fieldVersionId"
        ],
        "type": "object",
        "properties": {
          "fieldVersionId": {
            "type": "string",
            "description": "ID of the item field",
            "format": "uuid"
          },
          "strValue": {
            "type": "string",
            "description": "Value of the item field, depending on the type one of these will be filled",
            "nullable": true
          },
          "intValue": {
            "type": "integer",
            "description": "Value of the item field, depending on the type one of these will be filled",
            "format": "int64",
            "nullable": true
          },
          "decimalValue": {
            "type": "number",
            "description": "Value of the item field, depending on the type one of these will be filled",
            "format": "double",
            "nullable": true
          },
          "dateTimeValue": {
            "type": "string",
            "description": "Value of the item field, depending on the type one of these will be filled",
            "format": "date-time",
            "nullable": true
          },
          "boolValue": {
            "type": "boolean",
            "description": "Value of the item field, depending on the type one of these will be filled",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Model for the values of an item"
      },
...
      "StringStringListKeyValuePair": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}